Break out legacy ValuesResolver code
